backend/requirements.txt
fastapi==0.115.0
email-validator==2.2.0
uvicorn[standard]==0.30.6
pydantic-settings==2.5.2
asyncpraw==7.8.0
python-dotenv==1.0.1
google-genai==0.3.0
supabase==2.15.0
redis==5.2.1
pytest==8.3.5
